[Design] 토스티파이 루트에 적용
