+++ client/src/routes/Myinfo/MyinfoPage.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import styled, { css } from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-import { ToastContainer, toast } from 'react-toastify';
+import { toast } from 'react-toastify';
 import { BsFillExclamationCircleFill } from 'react-icons/bs';
 import customAxios from '../../api/customAxios';
 import { userInfo } from '../../recoil/userAtom';
@@ -32,13 +32,14 @@ function MyinfoPage() {
     const nickRegEx = /^[가-힣a-z0-9_-]{2,20}$/;
     if (!nickRegEx.test(nickname)) {
       toast.error(
-        '닉네임은 2자이상 20자 이하, 한글, 영문 대소문자, 숫자만 사용할 수 있습니다.\n \n 아무것도 입력하지 않은 경우, 카카오톡 프로필 이름이 닉네임으로 설정됩니다.',
+        '닉네임은 최소 2자 이상 최대 20자 이하, \n 한글, 영문 대소문자와 숫자만 사용할 수 있습니다.\n \n 아무것도 입력하지 않은 경우, 카카오톡 프로필 이름이 닉네임으로 설정됩니다.',
         {
+          autoClose: 5000,
           theme: 'dark',
           style: {
             fontSize: '1.4rem',
-            lineHeight: '1.6rem',
-            fontWeight: '500',
+            lineHeight: '2rem',
+            fontWeight: '400',
           },
         },
       );
@@ -161,7 +162,7 @@ function MyinfoPage() {
           onClick={handleSubmit}
         />
       </SettingForm>
-      <StyledToastContainer
+      {/* <StyledToastContainer
         position="top-right"
         autoClose={5000}
         hideProgressBar={false}
@@ -172,9 +173,7 @@ function MyinfoPage() {
         draggable
         pauseOnHover
         theme="dark"
-      />
-
-      <StyledToastContainer />
+      /> */}
     </SettingPageConainer>
   );
 }
@@ -322,14 +321,14 @@ const SubmitButton = styled(Button)`
   }
 `;
 
-const StyledToastContainer = styled(ToastContainer)`
-  &&&.Toastify__toast-container {
-    margin-top: 8rem;
-  }
+// const StyledToastContainer = styled(ToastContainer)`
+//   &&&.Toastify__toast-container {
+//     margin-top: 8rem;
+//   }
 
-  .Toastify__toast-body {
-    white-space: pre-line;
-  }
-`;
+//   .Toastify__toast-body {
+//     white-space: pre-line;
+//   }
+// `;
 
 export default MyinfoPage;
