+++ client/src/routes/Root.jsx
@@ -2,6 +2,8 @@ import { useEffect } from 'react';
 import styled from 'styled-components';
 import { Outlet, useNavigation } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
+import { ToastContainer } from 'react-toastify';
+
 import { userState } from '../recoil/userAtom';
 import Header from '../components/Header';
 import Footer from '../components/Footer';
@@ -30,6 +32,18 @@ function Root() {
       </Container>
       <Footer />
       <ScrollToTop />
+      <StyledToastContainer
+        position="top-right"
+        autoClose={2000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="dark"
+      />
     </>
   );
 }
@@ -39,6 +53,16 @@ const Container = styled.div`
   max-width: 1120px;
   min-height: 75vh;
   margin: auto;
+`;
+
+const StyledToastContainer = styled(ToastContainer)`
+  &&&.Toastify__toast-container {
+    margin-top: 8rem;
+  }
+
+  .Toastify__toast-body {
+    white-space: pre-line;
+  }
 `;
 
 export default Root;
